% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_model.R
\name{cache_model}
\alias{cache_model}
\title{Function to Save a Generic H2o Model Object Using a Saved Model Name As a Reference}
\usage{
cache_model(object = NULL, model_name, model_path = "../data-raw/models")
}
\arguments{
\item{object}{A h2o model object. If passed as \code{NULL} then the function will attempt to load a previously
saved model. Defaults to \code{NULL}.}

\item{model_name}{A character string indicating the model name, this will be saved in the \code{model_names}
subdirectory as an .rds object containing the h2o specified model name.}

\item{model_path}{A character string indicating the path into which to save the model and the model name.}
}
\value{
A h2o model object
}
\description{
This function uses a saved model name as a reference in order to find a h2o models
file path and then loads the h2o model using the inbuilt h2o functions. This function depends on
\code{\link[h2ohelper]{load_model}} in order to load in a saved h20 model object.
}
\examples{

}
\seealso{
load_model
}
