% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Function to Load a Generic H2o Model Object Using a Saved Model Name}
\usage{
load_model(model_name, model_path = NULL, model_name_path = NULL)
}
\arguments{
\item{model_name}{A character string containing the name of an .rds object
that contains the name of the model to load}

\item{model_path}{A character string containing the file path for folder
containing the models.}

\item{model_name_path}{A character string containing the file path for folder
containing the model names}
}
\value{
A h2o model object
}
\description{
Function to Load a Generic H2o Model Object Using a Saved Model Name
}
\examples{

}
